# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ai Pop Cute Support Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, unread, mentions, messages, activity]
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark single notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
  /notifications/{notificationId}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /conversations:
    get:
      operationId: getConversations
      tags: [messages]
      summary: Get all conversations
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationListResponse"
    post:
      operationId: createConversation
      tags: [messages]
      summary: Start a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{conversationId}:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Get a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{conversationId}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /conversations/{conversationId}/seen:
    post:
      operationId: markConversationSeen
      tags: [messages]
      summary: Mark conversation messages as seen
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as seen
  /conversations/group:
    post:
      operationId: createGroupConversation
      tags: [messages]
      summary: Create a group chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  # ── Search ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, numericId, name, username, isVerified, role, followersCount, followingCount, postsCount, likesCount, createdAt]
      properties:
        id:
          type: integer
        numericId:
          type: integer
        name:
          type: string
        username:
          type: string
        bio:
          type: ["string", "null"]
        alternateBio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        role:
          type: string
          enum: [user, creator]
        isPrivate:
          type: boolean
        mixChannelId:
          type: ["string", "null"]
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        likesCount:
          type: integer
        totalVotes:
          type: integer
        profileViews:
          type: integer
        isFollowing:
          type: ["boolean", "null"]
        isFollowRequestSent:
          type: ["boolean", "null"]
        isBlocked:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
...
    Notification:
      type: object
      required: [id, userId, type, title, body, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: string
        relatedId:
          type: ["integer", "null"]
        relatedType:
          type: ["string", "null"]
        actorId:
          type: ["integer", "null"]
        actor:
          $ref: "#/components/schemas/UserProfile"
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationListResponse:
      type: object
      required: [notifications, total, page, unreadCount]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        page:
          type: integer
        unreadCount:
          type: integer
...
    Conversation:
      type: object
      required: [id, isGroup, createdAt]
      properties:
        id:
          type: integer
        isGroup:
          type: boolean
        groupName:
          type: ["string", "null"]
        groupAvatarUrl:
          type: ["string", "null"]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ConversationListResponse:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
    ConversationInput:
      type: object
      required: [participantId]
      properties:
        participantId:
          type: integer
    GroupInput:
      type: object
      required: [groupName, memberIds]
      properties:
        groupName:
          type: string
          minLength: 1
        groupAvatarUrl:
          type: ["string", "null"]
        memberIds:
          type: array
          items:
            type: integer
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/UserProfile"
        content:
          type: string
        mediaUrl:
          type: ["string", "null"]
        replyToId:
          type: ["integer", "null"]
        isSeen:
          type: boolean
        isDelivered:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        mediaUrl:
          type: ["string", "null"]
        replyToId:
          type: ["integer", "null"]
    MessageListResponse:
      type: object
      required: [messages, total, page]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        total:
          type: integer
        page:
          type: integer

    # ── Search Schemas ───────────────────────────────────────────────────────
...